ConstrainedDFO.jl

This documentation is a work in progress and is by no means reliable yet!

ConstrainedDFO.AbstractInvertibilityBound — Type

AbstractInvertibilityBound

A formula to compute a lower bound on the invertibility_radius of a manifold.

ConstrainedDFO.EqualityManifold — Type

EqualityManifold <: AbstractManifold{ℝ}

A smooth Riemannian submanifold of $\mathbb{R}^n$ defined as the set

\[ \mathcal{M}=\left\{x\in\mathbb{R}^n : h(x)=0\right\}\]

for some smooth function $h: \mathbb{R}^n\to\mathbb{R}$ such that $\nabla h(x)$ has full rank for all $x\in\mathcal{M}$.

Fields

defining_function: the function $h$ as described above.

dimension: the dimension of the manifold, defined as the common dimension of its tangent spaces.

ConstrainedDFO.NOverSpectral — Type

NOverSpectral <: AbstractInvertibilityBound

Computes a lower bound on the invertibility_radius of the ProjectionRetraction as

\[ \frac{n}{\max\{\lambda(H_{h_i}(x)) : i\in\{1,...,p\}\}}.\]

with $\lambda(H_{h_i})$ the spectral radius of the Hessian matrix of the defining subfunction $h_i$ for M.

ConstrainedDFO.NOverSqrtSpectral — Type

NOverSqrtSpectral <: AbstractInvertibilityBound

Computes a lower bound on the invertibility_radius of the ProjectionRetraction as

\[ \frac{n}{\sqrt{\max\{\lambda(H_{h_i}(x)) : i\in\{1,...,p\}\}}}.\]

with $\lambda(H_{h_i})$ the spectral radius of the Hessian matrix of the defining subfunction $h_i$ for M.

ConstrainedDFO.OneOverSpectral — Type

OneOverSpectral <: AbstractInvertibilityBound

Computes a lower bound on the invertibility_radius of the ProjectionRetraction as

\[ \frac{1}{\max\{\lambda(H_{h_i}(x)) : i\in\{1,...,p\}\}}.\]

with $\lambda(H_{h_i})$ the spectral radius of the Hessian matrix of the defining subfunction $h_i$ for M.

ConstrainedDFO.OneOverSqrtSpectral — Type

OneOverSqrtSpectral <: AbstractInvertibilityBound

Computes a lower bound on the invertibility_radius of the ProjectionRetraction as

\[ \frac{1}{\sqrt{\max\{\lambda(H_{h_i}(x)) : i\in\{1,...,p\}\}}}.\]

with $\lambda(H_{h_i})$ the spectral radius of the Hessian matrix of the defining subfunction $h_i$ for M.

ConstrainedDFO.RDFOState — Type

RDFOState <: AbstractManoptSolverState

AbstractManoptSolverState dedicated to the RDFO solver.

Fields

p is the current iterate on the manifold.

d is the current best tangent vector found at $T_p\mathcal{M}$.

ConstrainedDFO.StopAfterEvaluation — Type

StopAfterEvaluation <: DFStoppingCriterion

A functor for a stopping criterion to stop after a maximal number of blackbox evaluations. Fields and constructor are the same as StopAfterIteration.

ConstrainedDFO.get_eval_budget — Method

get_eval_budget(em::AbstractEvalManager)

ConstrainedDFO.invertibility_radius — Method

invertibility_bound(M::EqualityManifold, p; m::AbstractRetractionMethod, b::AbstractInvertibilityBound)

Return a lower bound on the invertibility_radius of the EqualityManifold M at p, endowed with AbstractRetractionMethod m. The bound is computed according to the formula given by the AbstractInvertibilityBound b.

ConstrainedDFO.invertibility_radius — Method

invertibility_radius(M::AbstractManifold, p; m::AbstractRetractionMethod, b::AbstractInvertibilityBound)

Return a lower bound on the invertibility_radius` of M at p when endowed with the AbstractRetractionMethod m. The chosen bound is computed with the formula contained in b, and can be the exact value of the invertibility radius.

ConstrainedDFO.latin_hypercube_sampling — Method

latin_hypercube_sampling(n, lb, ub, p)

Perform a latin hypercube sampling with given bounds, and return p sampled points. It is assumed that bounds are well-defined.

ConstrainedDFO.solve_problem — Method

solve_problem(::AbstractDFSolver, ::AbstractBenchmarkProblem)

ConstrainedDFO.update_remaining_evals! — Method

update_remaining_evals(dem::DefaultEvalManager)